Fix. #113

+++ morphir/sdk/core/src/morphir/sdk/LocalTime.scala
@@ -7,13 +7,13 @@ object LocalTime {
   type LocalTime = java.time.LocalTime
 
   def addHours(hours: Basics.Int)(localTime: LocalTime): LocalTime =
-    localTime.plusHours(hours)
+    localTime.plusHours(hours.toLong)
 
   def addMinute(minutes: Basics.Int)(localTime: LocalTime): LocalTime =
-    localTime.plusMinutes(minutes)
+    localTime.plusMinutes(minutes.toLong)
 
   def addSeconds(seconds: Basics.Int)(localTime: LocalTime): LocalTime =
-    localTime.plusSeconds(seconds)
+    localTime.plusSeconds(seconds.toLong)
 
   def diffInHours(localTime1: LocalTime)(localTime2: LocalTime): Basics.Int =
     ChronoUnit.HOURS.between(localTime1, localTime2).toInt
